# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user detail (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot delete self
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/projects/{id}/stop:
    post:
      operationId: adminStopProject
      tags: [admin]
      summary: Stop any user's project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /admin/projects/{id}:
    delete:
      operationId: adminDeleteProject
      tags: [admin]
      summary: Delete any user's project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/users/{id}/role:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user's role and/or plan (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/credits:
    post:
      operationId: adminAdjustCredits
      tags: [admin]
      summary: Manually adjust a user's credit balance (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCreditsInput"
      responses:
        "200":
          description: Adjusted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/usage:
    get:
      operationId: adminGetUsage
      tags: [admin]
      summary: Get AI-proxy usage analytics (admin only)
      parameters:
        - name: days
          in: query
          required: false
          description: Rentang waktu ke belakang dalam hari (1–365). Default 30.
          schema:
            type: integer
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: Usage analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsage"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # Auth schemas
...
    Project:
      type: object
      required: [id, userId, name, runtime, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        repoUrl:
          type: ["string", "null"]
        runtime:
          type: string
          enum: [nodejs, python, php, static]
        status:
          type: string
          enum: [idle, running, stopped, building, deploying, failed]
        domain:
          type: ["string", "null"]
        createdAt:
          type: string
        lastDeployedAt:
          type: ["string", "null"]
...
    UserWithStats:
      type: object
      required: [id, email, name, role, plan, credits, createdAt, projectCount]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        plan:
          type: string
          enum: [hobby, pro, team]
        credits:
          type: integer
        createdAt:
          type: string
        projectCount:
          type: integer
        lastLoginAt:
          type: ["string", "null"]
    AdminUpdateUserInput:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [hobby, pro, team]
    AdjustCreditsInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          description: Jumlah kredit yang ditambah (positif) atau dikurangi (negatif).
        note:
          type: string
          description: Catatan alasan penyesuaian.
...
    AdminUsageTotals:
      type: object
      required: [requests, totalTokens, promptTokens, completionTokens, credits]
      properties:
        requests:
          type: integer
        totalTokens:
          type: integer
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        credits:
          type: integer
    AdminUsageByModel:
      type: object
      required: [model, requests, totalTokens, credits]
      properties:
        model:
          type: string
        requests:
          type: integer
        totalTokens:
          type: integer
        credits:
          type: integer
    AdminUsageByUser:
      type: object
      required: [userId, email, name, requests, totalTokens, credits]
      properties:
        userId:
          type: integer
        email:
          type: string
        name:
          type: string
        requests:
          type: integer
        totalTokens:
          type: integer
        credits:
          type: integer
    AdminUsageDaily:
      type: object
      required: [day, requests, totalTokens, credits]
      properties:
        day:
          type: string
          description: Tanggal dalam format YYYY-MM-DD.
        requests:
          type: integer
        totalTokens:
          type: integer
        credits:
          type: integer
    AdminUsage:
      type: object
      required: [rangeDays, since, totals, byModel, topUsers, daily]
      properties:
        rangeDays:
          type: integer
        since:
          type: string
          format: date-time
        totals:
          $ref: "#/components/schemas/AdminUsageTotals"
        byModel:
          type: array
          items:
            $ref: "#/components/schemas/AdminUsageByModel"
        topUsers:
          type: array
          items:
            $ref: "#/components/schemas/AdminUsageByUser"
        daily:
          type: array
          items:
            $ref: "#/components/schemas/AdminUsageDaily"